64-bit output fix
date	2002.11.15.16.48.00;	author rmoore1;	state Exp;

+++ source/components/utilities/utobject.c
@@ -1,7 +1,7 @@
 /******************************************************************************
  *
  * Module Name: utobject - ACPI object create/delete/size/cache routines
- *              $Revision: 1.78 $
+ *              $Revision: 1.79 $
  *
  *****************************************************************************/
 
@@ -246,8 +246,8 @@ AcpiUtCreateBufferObject (
     Buffer = ACPI_MEM_CALLOCATE (BufferSize);
     if (!Buffer)
     {
-        ACPI_REPORT_ERROR
-            (("CreateBuffer: could not allocate size %X\n", BufferSize));
+        ACPI_REPORT_ERROR (("CreateBuffer: could not allocate size %X\n", 
+            (UINT32) BufferSize));
         AcpiUtRemoveReference (BufferDesc);
         return_PTR (NULL);
     }
